fix HelloDynamo exercise 7
