+++ CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/1-Make-Nancy-Webserver/NancyServer/IndexModule.cs
@@ -1,0 +1,16 @@
+﻿using Nancy;
+using System;
+
+namespace NancyServer
+{
+    public class IndexModule : NancyModule
+    {
+        public IndexModule() : base("/")
+        {
+            Get["/"] = x => {
+                return "Hello, the server is now running. </br>" +
+                "It is now " + DateTime.Now.ToLongTimeString();
+            };
+        }
+    }
+}
