--- CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/1-Make-Nancy-Webserver/Server/Server/Bootstrapper.cs
@@ -1,20 +1,0 @@
-﻿using Nancy;
-using Nancy.Diagnostics;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DynamoServer.Server
-{
-    public class Bootstrapper : DefaultNancyBootstrapper
-    {
-        private const string _password = "dynamo";
-
-        protected override DiagnosticsConfiguration DiagnosticsConfiguration
-        {
-            get { return new DiagnosticsConfiguration { Password = _password }; }
-        }
-    }
-}
