--- CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/1-Make-Nancy-Webserver/Server/Views/index.html
@@ -1,17 +1,0 @@
-﻿<!DOCTYPE html>
-
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml">
-<head>
-    <meta charset="utf-8" />
-    <title>Welcome to Dynamo Server.</title>
-</head>
-<body>
-    <p>The server is now running, feel free to explore the endpoints :</p>
-    <ul>
-        <li><a href="http://localhost:1234/time">/time</a></li>
-        <li><a href="http://localhost:1234/hello">/hello</a></li>
-        <li><a href="http://localhost:1234/hello/AU-Workshop">/hello/{name}</a></li>
-        <li><a href="http://localhost:1234/echo/7">/echo/{value:int}</a></li>
-    </ul>
-</body>
-</html>
