+++ CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/2-Webserver-Library/Server/Server.csproj
@@ -50,7 +50,6 @@
   <ItemGroup>
     <Compile Include="Modules\IndexModule.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Server\Bootstrapper.cs" />
     <Compile Include="Server\DynamoWebServer.cs" />
   </ItemGroup>
   <ItemGroup>
