--- CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/3-Reference-Dynamo-Server/Server/Views/test.txt
@@ -1,1 +1,0 @@
-hello from notepad
