--- CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/3-Reference-Dynamo-Server/Server/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Nancy" version="1.4.5" targetFramework="net452" />
-  <package id="Nancy.Hosting.Self" version="1.4.1" targetFramework="net452" />
-</packages>
