+++ CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/3-Reference/Extensions/Events.cs
@@ -44,10 +44,5 @@ namespace DynamoServer.Extensions
         {
             await ServerViewExtension.StopServerAsync();
         }
-
-        public static void OnCheckServerStatus(object sender, RoutedEventArgs e)
-        {
-            MessageBox.Show(ServerViewExtension.GetServerStatus());
-        }
     }
 }
