+++ CBW227911 - Control Dynamo from the Web/Exercises/DynamoServer/3-Reference/Server/Server.csproj
