+++ CBW227911 - Control Dynamo from the Web/Exercises/HelloDynamo/3-Hello-ViewExtension/HelloDynamo/ExtensionExample.cs
@@ -1,4 +1,5 @@
-﻿using System.Windows;
+﻿using Dynamo.Extensions;
+using System.Windows;
 
 namespace HelloDynamo
 {
